Move MCP tools to subfolder
