+++ scripts/task-mcp/tools/find_app_icon.py
